Fixing includes and randgen (#50)
* fixed includes for all files

* updated unit to use double instead of int

* fixed double health and added output file

* updated gitignore

* removed output file

* Update EarthArmy.cpp

---------

+++ BattleUnits/AlienDrone.cpp
@@ -1,8 +1,11 @@
 #include "AlienDrone.h"
-
 #include "../Game.h"
+#include "../DataStructures/LinkedListStack.h"
+#include "../DataStructures/LinkedQueue.h"
 
-AlienDrone::AlienDrone(Game* game, int joinTime, int health, int power, int attackCapacity)
+
+
+AlienDrone::AlienDrone(Game* game, int joinTime, double health, double power, int attackCapacity)
     : Unit(game,AD,joinTime, health, power, attackCapacity) {
 }
