+++ BattleUnits/AlienMonster.cpp
@@ -1,8 +1,10 @@
 #include "AlienMonster.h"
-
 #include "../Game.h"
+#include "../DataStructures/LinkedQueue.h"
+#include "../DataStructures/LinkedListStack.h"
 
-AlienMonster::AlienMonster(Game* game, int joinTime, int health, int power, int attackCapacity)
+
+AlienMonster::AlienMonster(Game* game, int joinTime, double health, double power, int attackCapacity)
     : Unit(game,AM, joinTime, health, power, attackCapacity) {
 }
