+++ BattleUnits/AlienMonster.h
@@ -1,14 +1,11 @@
 #ifndef ALIEN_MONSTER_H
 #define ALIEN_MONSTER_H
 
-#include "../Armies/Army.h"
-#include "../DataStructures/LinkedListStack.h"
-#include "../DataStructures/LinkedQueue.h"
 #include "Unit.h"
 
 class AlienMonster : public Unit {
    public:
-    AlienMonster(Game* game, int joinTime, int health, int power, int attackCapacity);
+    AlienMonster(Game* game, int joinTime, double health, double power, int attackCapacity);
     void attack()override;
 };
