+++ BattleUnits/AlienSoldier.cpp
@@ -1,8 +1,9 @@
 #include "AlienSoldier.h"
-
 #include "../Game.h"
+#include "../DataStructures/LinkedQueue.h"
 
-AlienSoldier::AlienSoldier(Game* game, int joinTime, int health, int power, int attackCapacity)
+
+AlienSoldier::AlienSoldier(Game* game, int joinTime, double health, double power, int attackCapacity)
 	: Unit(game, AS, joinTime, health, power, attackCapacity) {
 }
