+++ BattleUnits/EarthSoldier.cpp
@@ -1,8 +1,9 @@
 #include "EarthSoldier.h"
-
 #include "../Game.h"
+#include "../DataStructures/LinkedQueue.h"
 
-EarthSoldier::EarthSoldier(Game* game, int joinTime, int health, int power, int attackCapacity)
+
+EarthSoldier::EarthSoldier(Game* game, int joinTime, double health, double power, int attackCapacity)
 	: Unit(game, ES, joinTime, health, power, attackCapacity) {
 }
