+++ BattleUnits/EarthTank.h
@@ -2,13 +2,12 @@
 #define EARTH_TANK_H
 
 #include "Unit.h"
-#include "../Game.h"
 
 class EarthTank : public Unit 
 {
 
 public:
-	EarthTank(Game* game, int joinTime, int health, int power, int attackCapacity);
+	EarthTank(Game* game, int joinTime, double health, double power, int attackCapacity);
 	void attack() override;
 };
