+++ BattleUnits/HealUnit.cpp
@@ -1,7 +1,9 @@
 #include "HealUnit.h"
 #include "../Game.h"
+#include "../DataStructures/LinkedQueue.h"
 
-HealUnit::HealUnit(Game* game, int joinTime, int health, int power, int attackCapacity)
+
+HealUnit::HealUnit(Game* game, int joinTime, double health, double power, int attackCapacity)
 	: Unit(game, ET, joinTime, health, power, attackCapacity)
 {
 }
